refactor: product detail page api to react-query
