+++ apps/web/src/app/products/[id]/components/PopularProudctsContainer.tsx
@@ -4,8 +4,13 @@ import { IProduct } from '@/graphql/interface';
 import React, { Suspense } from 'react';
 import PopularProducts from './PopularProudcts';
 import MoreProductsSkeleton from './MoreProductsSkeleton';
+import { ProductQuery } from '@/shared/api/gql/graphql';
 
-export default function PopularProductsContainer({ product }: { product: IProduct }) {
+export default function PopularProductsContainer({
+  product,
+}: {
+  product: NonNullable<ProductQuery['product']>;
+}) {
   return (
     <Suspense fallback={<MoreProductsSkeleton />}>
       <PopularProducts product={product} />
