+++ apps/web/src/entities/product/product.queries.ts
@@ -1,6 +1,9 @@
 import {
+  ProductGuidesQueryVariables,
+  ProductQueryVariables,
   QueryCommunityRandomRankingProductsQueryVariables,
   QueryProductsQueryVariables,
+  TogetherViewedProductsQueryVariables,
 } from '@/shared/api/gql/graphql';
 import { ProductService } from '@/shared/api/product';
 import { infiniteQueryOptions, queryOptions } from '@tanstack/react-query';
@@ -11,6 +14,17 @@ export const ProductQueries = {
     queryOptions({
       queryKey: [...ProductQueries.all(), 'ranking'],
       queryFn: () => ProductService.getRankingProducts(),
+    }),
+  product: (variables: ProductQueryVariables) =>
+    queryOptions({
+      queryKey: [
+        ...ProductQueries.all(),
+        'detail',
+        {
+          id: variables.id,
+        },
+      ],
+      queryFn: () => ProductService.getProduct(variables),
     }),
   products: (variables: QueryProductsQueryVariables) =>
     queryOptions({
@@ -52,5 +66,23 @@ export const ProductQueries = {
       getNextPageParam: (lastPage, allPages) => {
         return lastPage.products.at(-1)?.searchAfter?.[0];
       },
+    }),
+  productGuide: (variables: ProductGuidesQueryVariables) =>
+    queryOptions({
+      queryKey: [...ProductQueries.all(), 'guide', { productId: variables.productId }],
+      queryFn: () => ProductService.getProductGuides(variables),
+    }),
+
+  togetherViewed: (variables: TogetherViewedProductsQueryVariables) =>
+    queryOptions({
+      queryKey: [
+        ...ProductQueries.all(),
+        'viewed',
+        {
+          limit: variables.limit,
+          productId: variables.productId,
+        },
+      ],
+      queryFn: () => ProductService.getTogetherViewedProducts(variables),
     }),
 };
