+++ apps/web/src/shared/api/product/product.service.ts
@@ -1,13 +1,20 @@
 import { httpClient } from '@/shared/lib/http-client';
 import { graphql } from '../gql';
 import {
+  MutationCollectProductMutationVariables,
+  ProductGuidesQueryVariables,
+  ProductQueryVariables,
   QueryCommunityRandomRankingProductsQueryVariables,
   QueryProductsQueryVariables,
+  TogetherViewedProductsQueryVariables,
 } from '../gql/graphql';
 
 export class ProductService {
   static async getRankingProducts() {
     return httpClient.execute(QueryRankingProducts).then((res) => res.data);
+  }
+  static async getProduct(variables: ProductQueryVariables) {
+    return httpClient.execute(QueryProduct, variables).then((res) => res.data);
   }
   static async getProducts(variables: QueryProductsQueryVariables) {
     return httpClient.execute(QueryProducts, variables).then((res) => res.data);
@@ -19,6 +26,15 @@ export class ProductService {
       .execute(QueryCommunityRandomRankingProducts, variables)
       .then((res) => res.data);
   }
+  static async getProductGuides(variables: ProductGuidesQueryVariables) {
+    return httpClient.execute(QueryProductGuides, variables).then((res) => res.data);
+  }
+  static async getTogetherViewedProducts(variables: TogetherViewedProductsQueryVariables) {
+    return httpClient.execute(QueryTogetherViewedProducts, variables).then((res) => res.data);
+  }
+  static async collectProduct(variables: MutationCollectProductMutationVariables) {
+    return httpClient.execute(MutationCollectProduct, variables).then((res) => res.data);
+  }
 }
 
 const QueryRankingProducts = graphql(`
@@ -29,6 +45,52 @@ const QueryRankingProducts = graphql(`
       url
       price
       thumbnail
+    }
+  }
+`);
+
+const QueryProduct = graphql(`
+  query product($id: Int!) {
+    product(id: $id) {
+      id
+      providerId
+      category
+      categoryId
+      mallId
+      title
+      url
+      detailUrl
+      isHot
+      isEnd
+      isProfitUrl
+      price
+      postedAt
+      thumbnail
+      wishlistCount
+      positiveCommunityReactionCount
+      negativeCommunityReactionCount
+      provider {
+        id
+        name
+        nameKr
+        host
+      }
+      viewCount
+      mallName
+      guides {
+        id
+        title
+        content
+      }
+      prices {
+        id
+        target
+        type
+        price
+        createdAt
+      }
+      isMyWishlist
+      categoryName
     }
   }
 `);
@@ -98,5 +160,44 @@ const QueryCommunityRandomRankingProducts = graphql(`
       searchAfter
       postedAt
     }
+  }
+`);
+
+const QueryProductGuides = graphql(`
+  query productGuides($productId: Int!) {
+    productGuides(productId: $productId) {
+      id
+      title
+      content
+    }
+  }
+`);
+
+const QueryTogetherViewedProducts = graphql(`
+  query togetherViewedProducts($limit: Int!, $productId: Int!) {
+    togetherViewedProducts(limit: $limit, productId: $productId) {
+      id
+      title
+      mallId
+      url
+      isHot
+      isEnd
+      price
+      providerId
+      categoryId
+      category
+      thumbnail
+      provider {
+        nameKr
+      }
+      searchAfter
+      postedAt
+    }
+  }
+`);
+
+const MutationCollectProduct = graphql(`
+  mutation MutationCollectProduct($productId: Int!) {
+    collectProduct(productId: $productId)
   }
 `);
